+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Bugfixes
 
 - [#1938](https://github.com/influxdata/kapacitor/issues/1938): pagerduty2 should use routingKey rather than serviceKey
+- [#1982](https://github.com/influxdata/kapacitor/pull/1982): Fix KafkaTopic not working from TICKscript
 
 ## v1.5.0 [2018-05-17]